Decouple ONNX-GenAI test support
Move reusable graph, config, and package fixtures into a private non-collected support module. Rewire all consumers away from collected test modules and enforce the repository-wide import boundary without changing test collection.
