+++ src/mobius/integrations/onnx_genai/auto_export_test.py
@@ -17,6 +17,13 @@ import yaml
 from mobius._configs import QuantizationConfig
 from mobius._model_package import ModelPackage
 from mobius.integrations.onnx_genai import write_onnx_genai_config
+from mobius.integrations.onnx_genai._test_support import (
+    _Cfg,
+    _decoder_package,
+    _model,
+    _value,
+    _vlm_package,
+)
 from mobius.integrations.onnx_genai.auto_export import (
     _ddim_alpha_schedule,
     _flow_match_euler_schedule,
@@ -24,27 +31,11 @@ from mobius.integrations.onnx_genai.auto_export import (
     _looks_like_video_diffusion,
 )
 from mobius.integrations.onnx_genai.inference_metadata import SchedulerConfig
-from mobius.integrations.onnx_genai.inference_metadata_test import (
-    _decoder_model,
-    _model,
-    _value,
-)
 from mobius.integrations.onnx_genai.workflow_metadata import (
     HierarchicalAudioWorkflowConfig,
     build_decoder_workflow_metadata,
     build_hierarchical_audio_workflow_metadata,
 )
-
-
-@dataclasses.dataclass
-class _Cfg:
-    num_attention_heads: int = 16
-    num_key_value_heads: int = 4
-    head_dim: int = 64
-    hidden_size: int = 1024
-    max_position_embeddings: int = 8192
-    sliding_window: int | None = None
-    model_type: str = "qwen"
 
 
 @dataclasses.dataclass
@@ -160,67 +151,6 @@ def _video_diffusion_package() -> ModelPackage:
 
 class _MultimodalPkg(dict):
     config = _Cfg()
-
-
-@dataclasses.dataclass
-class _VisionCfg:
-    patch_size: int = 14
-    temporal_patch_size: int = 2
-    merge_size: int = 1
-    spatial_merge_size: int = 1
-    size: dict[str, int] = dataclasses.field(
-        default_factory=lambda: {"shortest_edge": 224, "longest_edge": 224}
-    )
-
-
-@dataclasses.dataclass
-class _VlmCfg(_Cfg):
-    vision: _VisionCfg = dataclasses.field(default_factory=_VisionCfg)
-    image_token_id: int = 32000
-    eos_token_id: int = 2
-
-
-def _vlm_package(*, audio: bool = False):
-    vision = _model(
-        "vision_encoder",
-        [
-            _value("pixel_values", ir.DataType.FLOAT, ["patches", 1176]),
-            _value("grid_thw", ir.DataType.INT64, ["images", 3]),
-        ],
-        [("image_features", ir.DataType.FLOAT, ["batch", 256, 32])],
-    )
-    embedding_inputs = [
-        _value("input_ids", ir.DataType.INT64, ["batch", "sequence"]),
-        _value("image_features", ir.DataType.FLOAT, ["batch", 256, 32]),
-    ]
-    components = {"vision_encoder": vision}
-    if audio:
-        components["audio_encoder"] = _model(
-            "audio_encoder",
-            [_value("input_features", ir.DataType.FLOAT, ["batch", 80, "frames"])],
-            [("audio_features", ir.DataType.FLOAT, ["batch", 64, 32])],
-        )
-        embedding_inputs.append(_value("audio_features", ir.DataType.FLOAT, ["batch", 64, 32]))
-    embedding = _model(
-        "embedding",
-        embedding_inputs,
-        [("inputs_embeds", ir.DataType.FLOAT, ["batch", "sequence", 32])],
-    )
-    decoder = _decoder_model(
-        [("inputs_embeds", ir.DataType.FLOAT, ["batch", "sequence", 32])],
-        position_shape=["batch", "sequence"],
-    )
-    components.update({"embedding": embedding, "decoder": decoder})
-    return ModelPackage(components, config=_VlmCfg())
-
-
-def _decoder_package(config=None):
-    model = _decoder_model(
-        [],
-        position_shape=["batch", "sequence"],
-        raw_token_input=True,
-    )
-    return ModelPackage({"model": model}, config=config or _Cfg())
 
 
 def test_dispatch_decoder(tmp_path):
